collecting data using nodes
this allows me to collect input, for how much nodes the user wants to collect, as well as collecting the input for each user node. this is useful because on this weeks assignment we will be implementing linked list still using the presidents secret arrival but instead of making various void functions and calling them, we can make a linked list to collect the arrays data, encode, decode and scramble the data for the array.